requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29

# file upload parsing
python-multipart>=0.0.9

# core RAG / LangChain deps
langchain>=0.1.19
langchain-community>=0.0.34
langchain-huggingface>=0.1.3
huggingface-hub>=0.23
torch>=2.2          # required by many HF models
faiss-cpu>=1.8

# document loaders / parsing
pypdf>=4.2          # for PyPDFLoader
unstructured>=0.14  # if you keep Unstructured loaders
pillow-heif>=0.15   # png/heif images inside PDFs (optional, remove if not needed)

# embeddings & transformers
sentence-transformers>=2.7
transformers>=4.52

# misc
python-dotenv>=1.0
